backhm/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
argcomplete==1.10.0
attrs==24.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.2
blis==1.0.1
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.10
dataclasses-json==0.6.7
docx2txt==0.8
EbookLib==0.17.1
extract-msg==0.23.1
fastapi==0.115.6
filetype==1.2.0
frozenlist==1.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.2.0
google-generativeai==0.8.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
IMAPClient==2.1.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10
langchain-google-genai==0.0.6
langchain-text-splitters==0.3.2
langcodes==3.5.0
langsmith==0.2.2
language_data==1.3.0
lxml==5.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
multidict==6.1.0
murmurhash==1.0.11
mypy-extensions==1.0.0
numpy==1.24.3
olefile==0.46
orjson==3.10.12
packaging==24.2
pandas==2.0.3
pdfminer.six==20181108
pillow==11.0.0
preshed==3.0.9
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
Pygments==2.18.0
PyMuPDF==1.25.0
pyparsing==3.2.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.19
python-pptx==0.6.18
pytz==2024.2
PyYAML==6.0.2
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
scikit-learn==1.3.0
scipy==1.14.1
setuptools==75.6.0
shellingham==1.5.4
six==1.17.0
smart-open==7.0.5
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soundfile==0.12.1
soupsieve==2.6
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.8.1
SQLAlchemy==2.0.36
srsly==2.5.0
starlette==0.41.3
tenacity==9.0.0
textract==1.6.3
thinc==8.3.2
threadpoolctl==3.5.0
tqdm==4.67.1
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==1.5.1
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.1
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.0
xlrd==1.2.0
XlsxWriter==3.2.0
yarl==1.18.3
email-validator==2.1.0
psycopg2-binary==2.9.9
